refactor: Redirect to liquidity token page after adding liquidity
